+++ spec/exvo_helpers/helpers_spec.rb
@@ -43,7 +43,6 @@ describe Exvo::Helpers do
     specify { Exvo::Helpers.music_uri.should match('music.exvo.com') }
     specify { Exvo::Helpers.pics_uri.should match('pics.exvo.com') }
     specify { Exvo::Helpers.preview_uri.should match('preview.exvo.com') }
-    specify { Exvo::Helpers.store_uri.should match('store.exvo.com') }
   end
 
   describe "host methods in production environment" do
@@ -59,7 +58,6 @@ describe Exvo::Helpers do
     specify { Exvo::Helpers.music_host.should eql('music.exvo.com') }
     specify { Exvo::Helpers.pics_host.should eql('pics.exvo.com') }
     specify { Exvo::Helpers.preview_host.should eql('preview.exvo.com') }
-    specify { Exvo::Helpers.store_host.should eql('store.exvo.com') }
   end
 
   describe "host methods in staging environment" do
@@ -79,7 +77,6 @@ describe Exvo::Helpers do
     specify { Exvo::Helpers.music_host.should eql('music.exvo.co') }
     specify { Exvo::Helpers.pics_host.should eql('pics.exvo.co') }
     specify { Exvo::Helpers.preview_host.should eql('preview.exvo.co') }
-    specify { Exvo::Helpers.store_host.should eql('store.exvo.co') }
   end
 
   describe "#auth_debug by default for production env" do
@@ -174,7 +171,6 @@ describe Exvo::Helpers do
     Exvo::Helpers.music_host = nil
     Exvo::Helpers.pics_host = nil
     Exvo::Helpers.preview_host = nil
-    Exvo::Helpers.store_host = nil
 
     ENV.delete("RACK_ENV")
     ENV.delete("CDN_HOST")